OCR-RAG-System-backend_branch/requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
pdf2image
opencv-python-headless
requests
firebase-admin
python-dotenv

python-multipart
langchain-google-genai
langchain-core
langchain-huggingface
pandas
numpy
apscheduler

# CPU-only versions to avoid CUDA dependencies
# Note: torch and torchvision are installed separately in Dockerfile
# to avoid pip hash mismatch errors with --index-url
easyocr
sentence-transformers
faiss-cpu
tiktoken


gspread
google-auth
google-auth-oauthlib
google-api-python-client
cryptography
openpyxl

# Authentication dependencies
PyJWT



